requirements.txt
atomicwrites==1.3.0
attrs==18.2.0
autoenv==1.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coverage==4.5.2
Flask==1.0.2
Flask-JWT-Extended==3.17.0
flask-marshmallow==0.9.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
marshmallow==2.18.0
more-itertools==5.0.0
pep8==1.7.1
pluggy==0.8.1
psycopg2-binary==2.7.7
py==1.7.0
pycodestyle==2.5.0
PyJWT==1.7.1
pytest==4.2.0
pytest-cov==2.6.1
python-coveralls==2.9.1
PyYAML==3.13
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
